Standardize and simplify stats animating script
Now to change the totals, we only need to change them in one place: just edit the HTML page itself (elements with the class `stats-animate` will automatically be animated).

Currently each new element will animate for 1 second longer than the last, as is the current behaviour. In future might change to add a data-animate-length attribute or similar if we need.