Fix edge-case networking bug
The custom override was not being applied correctly. In addition, when constructing the useReqCore we built the URL even if we didn't yet have the args. Both of these are now fixed.

Future work: We need tests here. We also need a better pattern for handling errors when manually calling useReq.