Wrap users panel in admin users page (#302)
If non-national team members attempted to access the users page, it made a request to the backend for the users. This caused the backend to fire off a 403 alert for us.

This PR fixes this behaviour by extracting the UsersPanel into its own component, and then wrapping it with RequireGroup, which prevents the rogue API request being sent.